chore(docs): add doctests for more evm modules
