+++ scripts/clippy
@@ -1,8 +1,8 @@
-cargo clippy --all-features -- --allow clippy::new_without_default --allow clippy::redundant_field_names --allow clippy::too_many_arguments --allow clippy::format_in_format_args --allow clippy::should_implement_trait
+cargo clippy --all-features -- --allow clippy::new_without_default --allow clippy::too_many_arguments --allow clippy::format_in_format_args --allow clippy::should_implement_trait
 
 # If --fix is used, run the above command with `--fix` instead of `--all-features`
 # to fix the code.
 
 if [ "$1" = "--fix" ]; then
-    cargo +nightly clippy --fix --all-features -Z unstable-options --allow-dirty --allow-staged -- --allow clippy::new_without_default --allow clippy::redundant_field_names --allow clippy::too_many_arguments --allow clippy::format_in_format_args --allow clippy::should_implement_trait
+    cargo +nightly clippy --fix --all-features -Z unstable-options --allow-dirty --allow-staged -- --allow clippy::new_without_default --allow clippy::too_many_arguments --allow clippy::format_in_format_args --allow clippy::should_implement_trait
 fi
